+++ nata-monkey/src/main/java/com/nata/monkeys/QLearningMonkey.java
@@ -14,16 +14,8 @@ import java.util.*;
  * Update: 2016-03-18 17:30
  */
 public class QLearningMonkey extends AbstractMonkey {
-    private Action restartAction = null;
-    private ActionFactory actionFactory = null;
-
     // <state,action,value>
     private Map<State, Map<Action, Double>> QMap = new HashMap<>();
-    private Set<String> activitySet = new HashSet<>();
-    private List<Action> actionList = new ArrayList<>();
-    private Action lastAction = null;
-    private Action backAction = null;
-    private Action homeAction = null;
 
     //final variables
     private final int ACTION_COUNTS = 200;
@@ -36,17 +28,13 @@ public class QLearningMonkey extends AbstractMonkey {
 
     public QLearningMonkey(String pkg, String act, AdbDevice device) {
         super("QLearningMonkey", pkg, act, device);
-        actionFactory = new ActionFactory(device);
-        restartAction = actionFactory.CreateRestartAction(getPkgAct());
-        backAction = actionFactory.createBackAction();
-        homeAction = actionFactory.createHomeAction();
     }
 
     private Action chooseActionFromState(State curState) {
         Map<Action, Double> actionTable = QMap.get(curState);
         if (actionTable == null || actionTable.size() == 0) {
             System.out.println("Error cannot get actionTable or no action can be choosed");
-            return backAction;
+            return getBackAction();
         }
 
         Action chosenAction;
@@ -110,30 +98,16 @@ public class QLearningMonkey extends AbstractMonkey {
         actionTable.put(action, reward + GAMA * future_reward);
     }
 
-
-    private void startApp() {
-        LogUtil.info("QLearning Monkey start playing...");
-        clearAppData();
-        LogUtil.info("App data cleaned!");
-        restartAction.fire();
-        actionList.add(restartAction);
-        LogUtil.info("Starting app success!");
-    }
-
     /**
-     * Add the current state to the value map of Q and add the new activity to activitySet
+     * Add the current state to the value map of Q
      * @return current state
      */
     @Override
     public State getCurrentState() {
         State state  = super.getCurrentState();
 
-        if (isInCurrentPkg()) {
-            activitySet.add(state.getActivity());
-        }
-
         if (QMap.get(state) == null) {
-            Map<Action, Double> table = actionFactory.getActionsFromState(state);
+            Map<Action, Double> table = getActionFactory().getActionsFromState(state);
             LogUtil.debug("ActionTable : " + table.toString());
             QMap.put(state, table);
             getDevice().screenShot(state.hashCode() + "");
@@ -142,47 +116,8 @@ public class QLearningMonkey extends AbstractMonkey {
         return state;
     }
 
-    /**
-     * help the monkey to get back to app
-     *
-     * @return the current state
-     */
-    private State getBackToApp() {
-        boolean forceQuit = false;
-        while (!isInCurrentPkg()) {
-            // if even the restart action cannot restart it ;
-            if(forceQuit){
-                clearAppData();
-                executeAction(restartAction);
-            }
-            else if(lastAction instanceof StartAppAction){
-                executeAction(homeAction);
-                executeAction(restartAction);
-                forceQuit = true;
-            }
-            // if monkey get out of pkg because of back action
-            else if (lastAction instanceof BackAction) {
-                executeAction(restartAction);
-            } else {
-                executeAction(backAction);
-                if (!isInCurrentPkg()) {
-                    executeAction(backAction);
-                }
-            }
-        }
 
-        return getCurrentState();
-    }
 
-    /**
-     * Fire the action and add the action to actionList
-     * @param action to be executed
-     */
-    private void executeAction(Action action){
-        action.fire();
-        actionList.add(action);
-        lastAction = action;
-    }
 
     @Override
     public void play() {
@@ -193,7 +128,8 @@ public class QLearningMonkey extends AbstractMonkey {
         int cnt = 0;
         while ((++cnt) < ACTION_COUNTS) {
             if (!isInCurrentPkg()) {
-                curState = getBackToApp();
+                getBackToApp();
+                curState =  getCurrentState();
                 continue;
             }
 
@@ -214,39 +150,15 @@ public class QLearningMonkey extends AbstractMonkey {
     }
 
     @Override
-    public void stop() {
-
-   }
-
-    @Override
     public void report() {
-        //report the activities reached
-        System.out.println("--------------------[Activities report]--------------------");
-        for (String activity : activitySet) {
-            System.out.println(activity);
+        super.report();
+
+        System.out.println("--------------------[State report]--------------------");
+        int stateCount = QMap.size();
+        System.out.println("State count: " + stateCount);
+        for (State state : QMap.keySet()) {
+            System.out.println(state);
         }
-
-//        System.out.println("--------------------[Action report]--------------------");
-//        int actionCount = actionList.size();
-//        System.out.println("Actions count: " + actionCount);
-//        for (Action action : actionList) {
-//            System.out.println(action);
-//        }
-
-        System.out.println("--------------------[Widget report]--------------------");
-        Set<Widget> widgetSet = getWidgetSet();
-        int widgetCount = widgetSet.size();
-        System.out.println("Widget count: " + widgetCount);
-        for (Widget node : widgetSet) {
-            System.out.println(node);
-        }
-
-//        System.out.println("--------------------[State report]--------------------");
-//        int stateCount = QMap.size();
-//        System.out.println("State count: " + stateCount);
-//        for (State state : QMap.keySet()) {
-//            System.out.println(state);
-//        }
     }
 
 }